Endringer etter QA
